requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic>=0.25.0
python-multipart==0.0.6
pdfplumber==0.10.3
PyPDF2==3.0.1
supabase>=2.15.0
google-cloud-aiplatform>=1.38.1
google-cloud-storage>=2.10.0
google-auth>=2.23.4
python-dotenv>=1.0.0
numpy
pydantic>=2.0.0
typing-extensions>=4.7.0
requests>=2.31.0
uuid==1.30 
